# aifo/slidescope/internal/server/public/openapi.yaml
basePath: /api/v1
definitions:
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.Mask:
    properties:
      maskHeight:
        description: Height of the mask in pixels
        type: integer
      maskId:
        description: Unique identifier for the mask
        type: string
      maskMpp:
        description: MPP of the mask
        type: number
      maskName:
        description: Display name for the mask
        type: string
      maskUri:
        description: URI to the mask file
        type: string
      maskWidth:
        description: Width of the mask in pixels
        type: integer
      slideId:
        description: Direct reference to slide ID
        type: string
      tilesUrl:
        description: URL to fetch mask tiles
        type: string
    type: object
  domain.MaskList:
    properties:
      masks:
        items:
          $ref: '#/definitions/domain.Mask'
        type: array
      slide_id:
        type: string
    type: object
...
  domain.User:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  handlers.UserInput:
    properties:
      password:
        example: password123
        type: string
      username:
        example: testuser
        type: string
    required:
      - password
      - username
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A digital pathology slide management and viewing API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SlideScope API
  version: "1.0"
paths:
...
  /api/v1/raster:
    get:
      description: Retrieve masks, optionally filtered by slide ID
      produces:
        - application/json
      responses:
        "200":
          description: List of masks
          schema:
            $ref: '#/definitions/domain.MaskList'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get masks
      tags:
        - masks
    post:
      consumes:
        - application/json
      description: Add a new mask and associate it with a slide
      parameters:
        - description: Mask information
          in: body
          name: mask
          required: true
          schema:
            $ref: '#/definitions/domain.Mask'
      produces:
        - application/json
      responses:
        "201":
          description: Created mask
          schema:
            $ref: '#/definitions/domain.Mask'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Slide not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Add a new mask
      tags:
        - masks
...
  /api/v1/slides/{slide_id}/raster:
    get:
      description: Retrieve masks, optionally filtered by slide ID
      parameters:
        - description: Slide ID to filter masks
          example: '"slide123"'
          in: path
          name: slide_id
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: List of masks
          schema:
            $ref: '#/definitions/domain.MaskList'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get masks
      tags:
        - masks
    post:
      consumes:
        - application/json
      description: Add a new mask and associate it with a slide
      parameters:
        - description: Slide ID (can be provided in URL or body)
          example: '"slide123"'
          in: path
          name: slide_id
          type: string
        - description: Mask information
          in: body
          name: mask
          required: true
          schema:
            $ref: '#/definitions/domain.Mask'
      produces:
        - application/json
      responses:
        "201":
          description: Created mask
          schema:
            $ref: '#/definitions/domain.Mask'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Slide not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Add a new mask
      tags:
        - masks
  /api/v1/slides/{slide_id}/raster/{mask_id}/tile/{z}/{x}/{y}.{format}:
    get:
      description: Retrieve a specific tile from a mask in XYZ tile format
      parameters:
        - description: Slide ID
          example: '"slide123"'
          in: path
          name: slide_id
          required: true
          type: string
        - description: Mask ID
          example: '"mask456"'
          in: path
          name: mask_id
          required: true
          type: string
        - description: Zoom level
          example: 0
          in: path
          name: z
          required: true
          type: integer
        - description: Tile X coordinate
          example: 0
          in: path
          name: x
          required: true
          type: integer
        - description: Tile Y coordinate
          example: 0
          in: path
          name: "y"
          required: true
          type: integer
        - description: Image format (only png supported)
          enum:
            - png
          example: '"png"'
          in: path
          name: format
          required: true
          type: string
      produces:
        - image/png
      responses:
        "200":
          description: Mask tile image data
          schema:
            type: file
        "400":
          description: Bad request - invalid parameters
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Mask, slide, or tile not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get mask tile
      tags:
        - masks
...
  /api/v1/users:
    post:
      consumes:
        - application/json
      description: Create a new user account with username and password
      parameters:
        - description: User credentials
          in: body
          name: user
          required: true
          schema:
            $ref: '#/definitions/handlers.UserInput'
      produces:
        - application/json
      responses:
        "201":
          description: Created user (password field will be empty)
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad request - invalid input
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Create a new user
      tags:
        - users
  /api/v1/users/{username}:
    get:
      description: Retrieve user information by username
      parameters:
        - description: Username
          example: '"testuser"'
          in: path
          name: username
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad request - username required
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
        - BearerAuth: []
      summary: Get user by username
      tags:
        - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"